Moved GOAudioGauge to the wxcontrols subdirectory
