+++ src/grandorgue/wxcontrols/GOAudioGauge.cpp
