+++ src/grandorgue/wxcontrols/GOAudioGauge.h
